gui_output/requirements.txt
flask>=2.3.0
requests>=2.31.0
numpy>=1.24.0
tensorflow>=2.13.0
biopython>=1.81